Update anomaly_detection.ipynb
I imported the two models that I am going to use to compare and see which would be the best for identifying threats using anomaly detection. I will be using the isolation forest algorithm and I will also be using AnoGAN which Stands for Anomaly Detection using Generative Adversarial Networks which means that the AnoGAN class uses a neural network which would be perfect to use because of the large amount of data.